refactor(runtimed): split peer presence module
